requirements.txt
fastapi==0.110.1
langchain~=0.1.14
langchain-google-genai==1.0.1
SQLAlchemy~=2.0.29
uvicorn~=0.29.0
pydantic~=2.6.4
passlib~=1.7.4
PyJWT~=2.8.0
python-multipart==0.0.9
